requirements.txt
streamlit>=1.30.0
gspread>=5.12.0
google-auth>=2.25.0
plotly>=5.18.0
pandas>=2.1.0
